requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
stripe>=5.0
django-environ>=0.9

# Dev / Testing
pytest>=7.0
pytest-django>=4.5
black>=23.0
flake8>=6.0
mypy>=1.0
python-dotenv>=1.0

